SMT-master/requirements.txt
fire
gin_config
lightning
loguru
musicdiff
numpy
opencv_python
opencv_python_headless
Pillow
rich
scikit-image
timm
#torch
torchinfo
#torchvision
transformers
visualizer
datasets
